fix(AI): fix test that failed on pipeline
