+++ .github/workflows/security.yml
@@ -24,14 +24,14 @@ jobs:
       - name: Lint (security, vulnerabilities)
         run: golangci-lint run --timeout 5m
 
-      - name: License compliance (Apache/MIT only)
+      - name: License compliance (Permissive OSI licenses)
         run: |
           go install github.com/google/go-licenses@latest
-          go-licenses check ./... --allowed_licenses=apache,mit
+          go-licenses check ./... --allowed_licenses=Apache-2.0,MIT,BSD-2-Clause,BSD-3-Clause,ISC
 
       - name: Check for findings and create issue if needed
         run: |
-          if grep -q 'forbidden license' <(go-licenses check ./... --allowed_licenses=apache,mit); then
+          if grep -q 'forbidden license' <(go-licenses check ./... --allowed_licenses=Apache-2.0,MIT,BSD-2-Clause,BSD-3-Clause,ISC); then
             gh issue create --title "License Compliance Issue" --body "Forbidden license detected in dependencies." --assignee ${{ github.repository_owner }}
           fi
         env:
