+++ cmd/gitversion-go/version_variables.go
@@ -1,14 +1,5 @@
 package main
 
-import (
-	"fmt"
-	"strings"
-
-	"gitversion-go/internal/gitversion"
-
-	"github.com/Masterminds/semver/v3"
-)
-
 // VersionVariables represents the version variables that can be used in the output.
 type VersionVariables struct {
 	Major         string `json:"Major"`
@@ -16,39 +7,4 @@ type VersionVariables struct {
 	Patch         string `json:"Patch"`
 	PreReleaseTag string `json:"PreReleaseTag"`
 	FullSemVer    string `json:"FullSemVer"`
-}
-
-func buildVersionVariables(version semver.Version, branchName string, commitsSinceTag int, config *gitversion.Config) VersionVariables {
-	finalVersion := version
-	matchingBranchConfig := config.GetBranchConfig(branchName)
-
-	if matchingBranchConfig != nil && commitsSinceTag > 0 {
-		tag := matchingBranchConfig.Tag
-		if tag == "use-branch-name" {
-			// Sanitize branch name for use in prerelease tag
-			sanitizedBranchName := strings.ReplaceAll(branchName, "/", "-")
-			tag = sanitizedBranchName
-		}
-
-		if tag != "" {
-			var prerelease string
-			if matchingBranchConfig.PreReleaseWeight > 0 {
-				prerelease = fmt.Sprintf("%s.%d.%d", tag, matchingBranchConfig.PreReleaseWeight, commitsSinceTag)
-			} else {
-				prerelease = fmt.Sprintf("%s.%d", tag, commitsSinceTag)
-			}
-			v, err := finalVersion.SetPrerelease(prerelease)
-			if err == nil {
-				finalVersion = v
-			}
-		}
-	}
-
-	return VersionVariables{
-		Major:         fmt.Sprintf("%d", finalVersion.Major()),
-		Minor:         fmt.Sprintf("%d", finalVersion.Minor()),
-		Patch:         fmt.Sprintf("%d", finalVersion.Patch()),
-		PreReleaseTag: finalVersion.Prerelease(),
-		FullSemVer:    finalVersion.String(),
-	}
 }
